"physical-lock": Interfaces, server and client implementations
This CL adds the interface for a lock server, an implementation
of a server object that exposes those interfaces, and a command-line
client for communicating with the server.

The server does not rely on any external service. It acts as an identity
provider and mounts itself in a local mounttable server that advertises
its location using MDNS.

Clients discover servers through the neighborhood. Clients are
authorized by a server only if they present a specific blessing rooted in the
server's identity provider. This blessing is handed out by the server at
the time of initialization.

A server can be initialized exactly once on a first-come-first-served
basis.

MultiPart: 1/2
Change-Id: I2cb4716b97fdd2036d4aef0a4b26b4e48d97b4be