Add support for circuitry using "real hardware" connected to a Pi.
This commit adds support for using the RaspberryPi's GPIO pins to
detect the state of the lock using a magnetic switch and change the
state of a lock using a relay (well, for now using a "buzzer").
See README.md for the circuit "diagram".

For testing on machines without the GPIO pins, a dummy implementation is
provided that prints status to STDOUT.

Currently, all this is placed in an "internal" package just so that
we can have a separate binary (hwtest) to test out the circuitry.
We might want to get rid of that binary and the "internal" package
altogether - folding the contents into "package main" for "lockd".

Change-Id: Ie8f3d72a8166ce81d80b6771a49591f2f3f17b4f