"physical-lock": Simplify scanning for lock devices
This CL simplifies the mechanism for discovering lock
devices by globbing the neighborhood.

Major changes
1) neighborhood-name under which a lock device's mounttable
is made visible is always of the form "lock-<lock name>", regardless of
whether the lock is claimed or unclaimed. Therefore discovering
lock devices simply requires a glob with the pattern "nh/lock-*".
This is a purely local operation at the clients. All locks objects
are accessible, by convention, under the name "nh/lock-<lock name>/lock.

2) The command for scanning lock objects globs over the neighborhood
in a loop in order to accomodate for any time lags in listening to MDNS
advertisments from the lock devices.

Minor change:
The command for scanning for lock objects is renamed from "find-locks"
to "scan".
Change-Id: Ie95b7dcffae7bc3a67de6b06b2e125a2359b09f4