"physical-lock": listkeys
This CL makes two changes:
1) Streamlines the naming of lock devices. Lock devices are always
named by the name using which they were claimed. This is also the
name that the device authenticates as and a prefix of every key for
the lock.
2) Adds 'listkeys' command to the command-line client that lists
the set of keys bound to the principal invoking the command and the
and the locks to the which they apply.

Change-Id: If3e8bbef4efdec910ebd69b7184478b26fbac229